feat: Not-simple Rule Types (#421)
Adding the "not-simple" rule types and their respective dataset
builders:
- [x] Dataset Metadata Check
- [x] Variable Metadata Check with Library
- [x] Value Check with Dataset Metadata
- [x] Value Check with Variable Metadata

There is also included in this PR a bug fix for rule 356, which amends
the check in exists_operator to also investigate the operation variables
for the target column first before probing deeper with _exists(). This
fix came about due to an error that was raised in regression after
implementing Value Check with Dataset Metadata.

I have intentionally left out the define xml specific rule types here as
we don't have the XML functionality to support adding them yet.

(regression not run for final commit yet)

---------

Co-authored-by: YetiGonk <maximo@verisian.com>